+++ deser-toml/src/document.rs
@@ -32,6 +32,8 @@ pub(crate) enum Value<'a> {
     /// An integer larger than `i64::MAX`.
     UInt(u64),
     Float(f64),
+    /// A float that is written as is (only used when serializing).
+    FloatText(Cow<'a, str>),
     Bool(bool),
     Datetime(Datetime),
     Table(usize),